requirements.txt
ollama==0.4.5
pyyaml==6.0.1